+++ R/mutate.R
@@ -56,7 +56,6 @@
 #'
 #' @export
 #' @inheritParams filter
-#' @inheritSection filter Tidy data
 #' @param ... <[`tidy-eval`][dplyr_tidy_eval]> Name-value pairs of expressions,
 #'   each with length 1 or the same length as the number of rows in the group
 #'   (if using [group_by()]) or in the entire input (if not using groups).
@@ -66,9 +65,25 @@
 #'   of the same name.
 #' @family single table verbs
 #' @return
-#' An object of the same type as `.data`. The rows will be left as; only
-#' the columns will be changed. `mutate()` preserves order of existing columns,
-#' adding new columns to the right. `transmute()` drops existing columns.
+#' An object of the same type as `.data`.
+#'
+#' For `mutate()`:
+#'
+#' * Rows are not affected.
+#' * Existing columns will be preserved unless explicitly modified.
+#' * New columns will be added to the right of existing columns.
+#' * Columns given value `NULL` will be removed
+#' * Groups will be recomputed if a grouping variable is mutated.
+#' * Data frame attributes are preserved.
+#'
+#' For `transmute()`:
+#'
+#' * Rows are not affected.
+#' * Apart from grouping variables, existing columns will be remove unless
+#'   explicitly kept.
+#' * Column order matches order of expressions.
+#' * Groups will be recomputed if a grouping variable is mutated.
+#' * Data frame attributes are preserved.
 #' @examples
 #' # Newly created variables are available immediately
 #' mtcars %>% as_tibble() %>% mutate(
@@ -133,33 +148,7 @@ mutate <- function(.data, ...) {
 #' @export
 mutate.data.frame <- function(.data, ...) {
   cols <- mutate_cols(.data, ...)
-  if (is.null(cols)) {
-    return(.data)
-  }
-
-  .data <- .data[setdiff(names(.data), cols$delete)]
-  .data[names(cols$add)] <- cols$add
-
-  .data
-}
-
-#' @export
-mutate.grouped_df <- function(.data, ...) {
-  cols <- mutate_cols(.data, ...)
-  if (is.null(cols)) {
-    return(.data)
-  }
-
-  data <- as_tibble(.data)
-  data <- data[setdiff(names(.data), cols$delete)]
-  data[names(cols$add)] <- cols$add
-
-  groups <- group_data(.data)
-  if (any(cols$change %in% names(groups))) {
-    grouped_df(data, group_vars(.data), group_by_drop_default(.data))
-  } else {
-    new_grouped_df(data, groups)
-  }
+  dplyr_col_modify(.data, cols)
 }
 
 #' @rdname mutate
@@ -171,33 +160,16 @@ transmute <- function(.data, ...) {
 #' @export
 transmute.data.frame <- function(.data, ...) {
   cols <- mutate_cols(.data, ...)
-  if (is.null(cols)) {
-    return(.data[integer()])
-  }
+  .data <- dplyr_col_modify(.data, cols)
 
-  .data <- .data[integer()]
-  .data[names(cols$add)] <- cols$add
+  out_cols <- c(
+    # ensure group vars present
+    setdiff(group_vars(.data), names(cols)),
+    # cols might contain NULLs
+    intersect(names(cols), names(.data))
+  )
 
-  .data
-}
-
-#' @export
-transmute.grouped_df <- function(.data, ...) {
-  cols <- mutate_cols(.data, ...)
-  if (is.null(cols)) {
-    return(.data[group_vars(.data)])
-  }
-
-  data <- as_tibble(.data)
-  data <- data[setdiff(group_vars(.data), names(cols$add))]
-  data[names(cols$add)] <- cols$add
-
-  groups <- group_data(.data)
-  if (any(cols$change %in% names(groups))) {
-    grouped_df(data, group_vars(.data), group_by_drop_default(.data))
-  } else {
-    new_grouped_df(data, groups)
-  }
+  .data[out_cols]
 }
 
 # Helpers -----------------------------------------------------------------
@@ -268,10 +240,6 @@ mutate_cols <- function(.data, ...) {
   }
 
   is_zap <- map_lgl(new_columns, inherits, "rlang_zap")
-
-  list(
-    add = new_columns[!is_zap],
-    delete = names(new_columns)[is_zap],
-    change = names(new_columns)
-  )
+  new_columns[is_zap] <- rep(list(NULL), sum(is_zap))
+  new_columns
 }