+++ R/summarise.R
@@ -33,20 +33,23 @@
 #'
 #' @export
 #' @inheritParams filter
-#' @inheritSection filter Tidy data
 #' @param ... <[`tidy-eval`][dplyr_tidy_eval]> Name-value pairs of summary
 #'   functions. The name will be the name of the variable in the result.
 #'   The value should be an expression that returns a single value like
 #'   `min(x)`, `n()`, or `sum(is.na(y))`.
 #' @family single table verbs
 #' @return
-#' An object of the same class as `.data`. It will contain one column for
-#' each grouping variable and each expression you supply. It will have
-#' one row for each combination of the grouping variables.
+#' An object _usually_ of the same type as `.data`.
 #'
-#' If `.data` is grouped, then the last group will be dropped,
-#' e.g.`df %>% group_by(x, y) %>% summarise(n())` will be grouped by
-#' `x`. This happens because each group now occupies only a single row.
+#' * The rows come from the underlying `group_keys()`.
+#' * The columns are a combination of the grouping keys and the summary
+#'   expressions that you provide.
+#' * If `x` is grouped by more than one variable, the output will be another
+#'   [grouped_df] with the right-most group removed.
+#' * If `x` is grouped by one variable, or is not grouped, the output will
+#'   be a [tibble].
+#' * Data frame attributes are **not** preserved, because `summarise()`
+#'   fundamentally creates a new data frame.
 #' @examples
 #' # A summary applied to ungrouped tbl returns a single row
 #' mtcars %>%
@@ -90,8 +93,8 @@ summarise.data.frame <- function(.data, ...) {
   if (!identical(cols$size, 1L)) {
     out <- vec_slice(out, rep(1:nrow(out), cols$size))
   }
-  out[names(cols$new)] <- cols$new
-  out
+
+  dplyr_col_modify(out, cols$new)
 }
 
 #' @export
@@ -99,8 +102,9 @@ summarise.grouped_df <- function(.data, ...) {
   out <- NextMethod()
 
   group_vars <- group_vars(.data)
-  if (length(group_vars) > 1) {
-    out <- grouped_df(out, group_vars[-length(group_vars)], group_by_drop_default(.data))
+  n <- length(group_vars)
+  if (n > 1) {
+    out <- grouped_df(out, group_vars[-n], group_by_drop_default(.data))
   }
 
   out